@@ -24,11 +24,13 @@
   },
   "dependencies": {
     "css-layout": "^1.0.0",
+    "prop-types": "^15.6.2",
     "wolfy87-eventemitter": "^4.2.11"
   },
   "devDependencies": {
     "babel-cli": "^6.23.0",
     "babel-core": "^6.23.0",
+    "babel-eslint": "^10.0.1",
     "babel-loader": "^6.3.2",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.23.0",
@@ -36,14 +38,15 @@
     "babel-preset-react": "^6.23.0",
     "babel-preset-stage-0": "^6.22.0",
     "babel-runtime": "^6.23.0",
+    "eslint": "^5.9.0",
     "eslint-plugin-mocha": "^4.8.0",
     "eslint-plugin-pnm": "^2.0.0",
     "eslint-plugin-pnm-frontend": "^1.0.2",
-    "eslint-plugin-react": "^6.10.0",
+    "eslint-plugin-react": "^7.11.1",
     "npm-run-all": "^4.0.2",
     "opn-cli": "^3.1.0",
-    "react": "^15.0.0",
-    "react-dom": "^15.0.0",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
     "rimraf": "^2.6.1",
     "webpack": "^2.2.1",
     "webpack-dev-server": "^2.4.1"
