@@ -12,6 +12,10 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _ = require('.');
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
@@ -31,11 +35,11 @@ const FlexContainer = props => {
 };
 
 FlexContainer.propTypes = {
-  layout: _react2.default.PropTypes.object.isRequired, // Passed via Layoutable.
-  id: _react2.default.PropTypes.string,
-  layoutRectProps: _react2.default.PropTypes.object,
-  renderLayoutRect: _react2.default.PropTypes.bool,
-  children: _react2.default.PropTypes.node
+  layout: _propTypes2.default.object.isRequired, // Passed via Layoutable.
+  id: _propTypes2.default.string,
+  layoutRectProps: _propTypes2.default.object,
+  renderLayoutRect: _propTypes2.default.bool,
+  children: _propTypes2.default.node
 };
 
 FlexContainer.defaultProps = {
