@@ -9,6 +9,10 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _wolfy87Eventemitter = require('wolfy87-eventemitter');
 
 var _wolfy87Eventemitter2 = _interopRequireDefault(_wolfy87Eventemitter);
@@ -114,11 +118,11 @@ exports.default = FlexContext;
 
 
 FlexContext.childContextTypes = {
-  styleTools: _react2.default.PropTypes.object.isRequired,
-  waitForLayoutCalculation: _react2.default.PropTypes.func.isRequired,
-  deregister: _react2.default.PropTypes.func.isRequired
+  styleTools: _propTypes2.default.object.isRequired,
+  waitForLayoutCalculation: _propTypes2.default.func.isRequired,
+  deregister: _propTypes2.default.func.isRequired
 };
 
 FlexContext.propTypes = {
-  children: _react2.default.PropTypes.node
+  children: _propTypes2.default.node
 };
