@@ -16,6 +16,10 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 // A utility component to assist in rendering a layout rectangle, useful for
@@ -46,7 +50,7 @@ const LayoutRect = props => {
 LayoutRect.defaultFormat = { fill: 'transparent' };
 
 LayoutRect.propTypes = {
-  layout: _react2.default.PropTypes.object.isRequired
+  layout: _propTypes2.default.object.isRequired
 };
 
 exports.default = LayoutRect;
