@@ -12,6 +12,10 @@ var _react = require('react');
 
 var _react2 = _interopRequireDefault(_react);
 
+var _propTypes = require('prop-types');
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
+
 var _layoutRect = require('./layout-rect');
 
 var _layoutRect2 = _interopRequireDefault(_layoutRect);
@@ -50,9 +54,7 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 // This used to be called FlexBox. The coupling between Layoutable and FlexBox
 // is largely unchanged from the original.
 
-/* eslint react/prop-types: "off" */
-
-const getDisplayName = Component => Component.displayName || Component.name || 'Component';
+const getDisplayName = Component => Component.displayName || Component.name || 'Component'; /* eslint react/prop-types: "off" */
 
 const Layoutable = (computeStyleFromProps = props => {}, options) => Composed => {
   const Wrapped = class extends _react2.default.Component {
@@ -134,13 +136,13 @@ const Layoutable = (computeStyleFromProps = props => {}, options) => Composed =>
   };
 
   Wrapped.contextTypes = {
-    styleTools: _react2.default.PropTypes.object.isRequired,
-    waitForLayoutCalculation: _react2.default.PropTypes.func.isRequired,
-    deregister: _react2.default.PropTypes.func.isRequired
+    styleTools: _propTypes2.default.object.isRequired,
+    waitForLayoutCalculation: _propTypes2.default.func.isRequired,
+    deregister: _propTypes2.default.func.isRequired
   };
 
   Wrapped.childContextTypes = {
-    styleTools: _react2.default.PropTypes.object.isRequired
+    styleTools: _propTypes2.default.object.isRequired
   };
 
   return Wrapped;
